Rename lib/ to internal/
