+++ package.json
@@ -40,7 +40,7 @@
     "playground-service-worker.js",
     "playground-typescript-worker.js",
     "playground-styles.{css,js,d.ts,d.ts.map}",
-    "lib/**/*.{js,d.ts,d.ts.map}",
+    "internal/**/*.{js,d.ts,d.ts.map}",
     "shared/**/*.{js,d.ts,d.ts.map}",
     "src/**/*.{ts,js}",
     "!src/test/**",
