+++ src/internal/version.ts
